Remove fileName argument from Hax_DumpActiveMemory
Print the Hax_DumpActiveMemory messages to stdout always.

This removes a dependency of Unix Core (Hax_TildeSubst) in
libhax.a. Print to stdout, rather than stderr as long term the
printing in libhax.a routines will be printing only to stdout.

This is a preparating for freestanding execution without
distinguished two output streams.