Fix use of stack after free
TraceVarProc() creates a dummy interpreter and evaluates it.

Set memoryp for dummy.

Initialize to an empty string dummy.result for interp->freeProc == 0.

After finishing evaluation, whenever the result is stored in
dummy.staticSpace, recreate it inside interp->staticSpace, rather
than pointing to the static buffer of dummy on the local stack.

This fixes MSan and ASan violation reports.